cfg_simplify: Fix phi node renaming corner case bug
Depending on the order of basic block, cfg_simplify! could occasionally
forget to renumber a phinode edge, causing at best a downstream assertion
and at worst a miscompile. The root of this issue is that ordinarily,
we detect whether a phinode edge requires later renaming by looking
at the last inserted index, and since we process all indices by order,
detecting whether we have already inserted a certain value is a simple
comparison. However, cfg_simplify! re-arranges the order of basic blocks,
so this check no longer works. Instead, we check whether the incoming
edge has already been scheduled. This is slightly less precise (because
it is possible for the edge itself to not have been scheduled, even
though the value has been), but correct and a slight imprecision here
at worst causes a negligible compile-time increase.